Move all of shell into one package
