+++ grakn-graql-shell/src/main/java/ai/grakn/graql/shell/License.java
@@ -16,7 +16,7 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.graql;
+package ai.grakn.graql.shell;
 
 import jline.console.ConsoleReader;
