+++ grakn-graql-shell/src/main/java/ai/grakn/graql/shell/ShellCommandCompleter.java
@@ -16,9 +16,8 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.graql.internal.shell;
+package ai.grakn.graql.shell;
 
-import ai.grakn.graql.GraqlShell;
 import jline.console.completer.Completer;
 
 import java.util.List;
